add navbar user with Avatar and DropdownMenu component
